Replace automation attempt with something simpler -- sqlite and openai for both image generation and transaltions
Original idea was to try to make use of my existing desktop for
generating these and have it always connected to a webserver. I'm going
to go a little bit simpler and do everything via CLI with a little bit
of manual CSV downloading. It'll be slightly more work, but if I only do
it ~1 time a week it really shouldn't matter too much.

And the development time is measured in hours instead of days, which
adds probably 20+ hours of studying time back to my project.